# Add NodeJs bindings benchmarks