I am going to try to add a couple of performance optimisation in NodeJs bindings, so it would be really useful.



<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1149.31<br/>plain u32: 1149.65<br/>aliased double: 1179.27<br/>plain double: 1190.37 | aliased u32: 1163.21<br/>plain u32: 1166.87<br/>aliased double: 1203.11<br/>plain double: 1211.36 |
| e2e_match_ch | Ops: 42.66 ± 0.25 ops/s. Best: 42.92 ops/s<br/>Total: 3071.30ms ± 18.80ms. Best: 3052.09ms<br/>Min time: 2.09ms ± 0.02ms<br/>Mean time: 23.44ms ± 0.15ms<br/>Median time: 15.89ms ± 0.13ms<br/>95th percentile: 80.46ms ± 0.96ms<br/>99th percentile: 97.67ms ± 0.33ms<br/>Max time: 102.14ms ± 0.57ms | Ops: 43.68 ± 0.06 ops/s. Best: 43.77 ops/s<br/>Total: 2999.34ms ± 4.33ms. Best: 2993.06ms<br/>Min time: 2.07ms ± 0.03ms<br/>Mean time: 22.90ms ± 0.03ms<br/>Median time: 15.77ms ± 0.08ms<br/>95th percentile: 75.73ms ± 0.23ms<br/>99th percentile: 91.39ms ± 0.23ms<br/>Max time: 100.04ms ± 0.33ms |
| e2e_match_mld | Ops: 62.37 ± 0.14 ops/s. Best: 62.65 ops/s<br/>Total: 2100.58ms ± 4.83ms. Best: 2090.95ms<br/>Min time: 1.82ms ± 0.02ms<br/>Mean time: 16.03ms ± 0.04ms<br/>Median time: 8.41ms ± 0.14ms<br/>95th percentile: 52.72ms ± 0.30ms<br/>99th percentile: 60.94ms ± 0.22ms<br/>Max time: 69.78ms ± 0.35ms | Ops: 61.85 ± 0.22 ops/s. Best: 62.22 ops/s<br/>Total: 2118.16ms ± 7.95ms. Best: 2105.57ms<br/>Min time: 1.80ms ± 0.02ms<br/>Mean time: 16.17ms ± 0.06ms<br/>Median time: 8.41ms ± 0.11ms<br/>95th percentile: 54.18ms ± 1.11ms<br/>99th percentile: 63.53ms ± 1.24ms<br/>Max time: 70.32ms ± 1.84ms |
| e2e_nearest_ch | Ops: 867.25 ± 5.37 ops/s. Best: 874.27 ops/s<br/>Total: 1152.97ms ± 7.36ms. Best: 1143.82ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.15ms ± 0.01ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.60ms ± 0.01ms<br/>99th percentile: 1.65ms ± 0.01ms<br/>Max time: 6.74ms ± 3.61ms | Ops: 866.78 ± 3.63 ops/s. Best: 873.80 ops/s<br/>Total: 1153.69ms ± 5.17ms. Best: 1144.43ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.15ms ± 0.01ms<br/>Median time: 1.07ms ± 0.00ms<br/>95th percentile: 1.60ms ± 0.01ms<br/>99th percentile: 1.65ms ± 0.01ms<br/>Max time: 5.74ms ± 2.64ms |
| e2e_nearest_mld | Ops: 865.66 ± 3.58 ops/s. Best: 871.46 ops/s<br/>Total: 1155.23ms ± 5.30ms. Best: 1147.50ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.16ms ± 0.01ms<br/>Median time: 1.07ms ± 0.00ms<br/>95th percentile: 1.60ms ± 0.01ms<br/>99th percentile: 1.66ms ± 0.01ms<br/>Max time: 5.86ms ± 2.71ms | Ops: 860.53 ± 11.68 ops/s. Best: 872.55 ops/s<br/>Total: 1162.62ms ± 16.10ms. Best: 1146.07ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.16ms ± 0.02ms<br/>Median time: 1.07ms ± 0.01ms<br/>95th percentile: 1.61ms ± 0.02ms<br/>99th percentile: 1.69ms ± 0.05ms<br/>Max time: 5.84ms ± 2.63ms |
| e2e_route_ch | Ops: 373.30 ± 1.42 ops/s. Best: 376.39 ops/s<br/>Total: 2679.03ms ± 10.66ms. Best: 2656.80ms<br/>Min time: 1.19ms ± 0.01ms<br/>Mean time: 2.68ms ± 0.01ms<br/>Median time: 2.69ms ± 0.00ms<br/>95th percentile: 3.52ms ± 0.02ms<br/>99th percentile: 3.95ms ± 0.03ms<br/>Max time: 6.74ms ± 2.39ms | Ops: 363.46 ± 5.07 ops/s. Best: 369.96 ops/s<br/>Total: 2753.04ms ± 40.13ms. Best: 2702.97ms<br/>Min time: 1.20ms ± 0.01ms<br/>Mean time: 2.75ms ± 0.04ms<br/>Median time: 2.77ms ± 0.04ms<br/>95th percentile: 3.65ms ± 0.07ms<br/>99th percentile: 4.06ms ± 0.07ms<br/>Max time: 6.78ms ± 2.27ms |
| e2e_route_mld | Ops: 307.67 ± 3.26 ops/s. Best: 310.63 ops/s<br/>Total: 3250.92ms ± 34.79ms. Best: 3219.27ms<br/>Min time: 1.19ms ± 0.01ms<br/>Mean time: 3.25ms ± 0.03ms<br/>Median time: 3.30ms ± 0.04ms<br/>95th percentile: 4.45ms ± 0.06ms<br/>99th percentile: 4.94ms ± 0.07ms<br/>Max time: 7.72ms ± 2.16ms | Ops: 309.41 ± 1.65 ops/s. Best: 312.60 ops/s<br/>Total: 3231.67ms ± 18.50ms. Best: 3198.93ms<br/>Min time: 1.17ms ± 0.01ms<br/>Mean time: 3.23ms ± 0.02ms<br/>Median time: 3.27ms ± 0.01ms<br/>95th percentile: 4.41ms ± 0.04ms<br/>99th percentile: 4.91ms ± 0.06ms<br/>Max time: 7.36ms ± 2.03ms |
| e2e_table_ch | Ops: 317.49 ± 0.61 ops/s. Best: 318.20 ops/s<br/>Total: 3149.66ms ± 6.65ms. Best: 3142.64ms<br/>Min time: 1.63ms ± 0.02ms<br/>Mean time: 3.15ms ± 0.01ms<br/>Median time: 3.15ms ± 0.01ms<br/>95th percentile: 4.37ms ± 0.03ms<br/>99th percentile: 4.72ms ± 0.04ms<br/>Max time: 8.18ms ± 2.21ms | Ops: 315.22 ± 0.58 ops/s. Best: 316.02 ops/s<br/>Total: 3172.27ms ± 5.87ms. Best: 3164.36ms<br/>Min time: 1.64ms ± 0.03ms<br/>Mean time: 3.17ms ± 0.01ms<br/>Median time: 3.17ms ± 0.01ms<br/>95th percentile: 4.40ms ± 0.03ms<br/>99th percentile: 4.73ms ± 0.06ms<br/>Max time: 8.30ms ± 2.46ms |
| e2e_table_mld | Ops: 108.90 ± 0.62 ops/s. Best: 109.61 ops/s<br/>Total: 9182.47ms ± 53.11ms. Best: 9123.65ms<br/>Min time: 3.67ms ± 0.09ms<br/>Mean time: 9.18ms ± 0.05ms<br/>Median time: 9.14ms ± 0.07ms<br/>95th percentile: 14.08ms ± 0.07ms<br/>99th percentile: 14.79ms ± 0.09ms<br/>Max time: 17.36ms ± 2.31ms | Ops: 109.90 ± 0.45 ops/s. Best: 110.49 ops/s<br/>Total: 9097.98ms ± 37.00ms. Best: 9050.91ms<br/>Min time: 3.68ms ± 0.06ms<br/>Mean time: 9.10ms ± 0.04ms<br/>Median time: 9.06ms ± 0.08ms<br/>95th percentile: 13.93ms ± 0.06ms<br/>99th percentile: 14.62ms ± 0.07ms<br/>Max time: 17.42ms ± 2.32ms |
| e2e_trip_ch | Ops: 96.84 ± 0.93 ops/s. Best: 98.83 ops/s<br/>Total: 10327.68ms ± 101.91ms. Best: 10118.12ms<br/>Min time: 1.55ms ± 0.03ms<br/>Mean time: 10.33ms ± 0.10ms<br/>Median time: 9.79ms ± 0.11ms<br/>95th percentile: 18.51ms ± 0.14ms<br/>99th percentile: 20.57ms ± 0.31ms<br/>Max time: 22.36ms ± 0.55ms | Ops: 100.58 ± 0.41 ops/s. Best: 101.10 ops/s<br/>Total: 9940.82ms ± 41.07ms. Best: 9891.51ms<br/>Min time: 1.43ms ± 0.12ms<br/>Mean time: 9.94ms ± 0.04ms<br/>Median time: 9.39ms ± 0.07ms<br/>95th percentile: 17.97ms ± 0.10ms<br/>99th percentile: 19.83ms ± 0.13ms<br/>Max time: 22.31ms ± 1.11ms |
| e2e_trip_mld | Ops: 59.26 ± 0.40 ops/s. Best: 59.69 ops/s<br/>Total: 16876.46ms ± 117.90ms. Best: 16752.12ms<br/>Min time: 1.64ms ± 0.28ms<br/>Mean time: 16.87ms ± 0.12ms<br/>Median time: 16.38ms ± 0.16ms<br/>95th percentile: 27.74ms ± 0.13ms<br/>99th percentile: 30.00ms ± 0.34ms<br/>Max time: 31.87ms ± 0.20ms | Ops: 58.83 ± 0.38 ops/s. Best: 59.30 ops/s<br/>Total: 17000.78ms ± 110.64ms. Best: 16864.57ms<br/>Min time: 1.58ms ± 0.24ms<br/>Mean time: 17.00ms ± 0.12ms<br/>Median time: 16.49ms ± 0.11ms<br/>95th percentile: 27.98ms ± 0.16ms<br/>99th percentile: 30.27ms ± 0.09ms<br/>Max time: 32.41ms ± 0.59ms |
| json-render | String: 5.43145ms<br/>Stringstream: 8.84586ms<br/>Vector: 6.47058ms | String: 5.4868ms<br/>Stringstream: 9.05192ms<br/>Vector: 6.52944ms |
| match_ch | Default radius: <br/>4.61265ms/req at 82 coordinate<br/>0.0562519ms/coordinate<br/>Radius 10m: <br/>16.1109ms/req at 82 coordinate<br/>0.196475ms/coordinate | Default radius: <br/>4.62303ms/req at 82 coordinate<br/>0.0563784ms/coordinate<br/>Radius 10m: <br/>16.1028ms/req at 82 coordinate<br/>0.196376ms/coordinate |
| match_mld | Default radius: <br/>3.14002ms/req at 82 coordinate<br/>0.0382929ms/coordinate<br/>Radius 10m: <br/>11.7684ms/req at 82 coordinate<br/>0.143517ms/coordinate | Default radius: <br/>2.94615ms/req at 82 coordinate<br/>0.0359286ms/coordinate<br/>Radius 10m: <br/>11.124ms/req at 82 coordinate<br/>0.135659ms/coordinate |
| node_match_ch | Ops: 283.8 ± 3.3 ops/s. Best: 286.8 ops/s | Ops: 285.7 ± 0.9 ops/s. Best: 286.8 ops/s |
| node_match_mld | Ops: 365.7 ± 2.5 ops/s. Best: 369.0 ops/s | Ops: 363.8 ± 2.4 ops/s. Best: 366.8 ops/s |
| node_nearest_ch | Ops: 11653.0 ± 529.9 ops/s. Best: 12688.7 ops/s | Ops: 11596.2 ± 664.1 ops/s. Best: 13016.9 ops/s |
| node_nearest_mld | Ops: 12040.3 ± 622.4 ops/s. Best: 13083.6 ops/s | Ops: 11900.0 ± 319.0 ops/s. Best: 12298.3 ops/s |
| node_route_ch | Ops: 1619.3 ± 21.1 ops/s. Best: 1667.8 ops/s | Ops: 1568.6 ± 28.3 ops/s. Best: 1606.9 ops/s |
| node_route_mld | Ops: 902.6 ± 11.6 ops/s. Best: 926.6 ops/s | Ops: 848.9 ± 16.9 ops/s. Best: 878.0 ops/s |
| node_table_ch | Ops: 249.1 ± 0.5 ops/s. Best: 249.7 ops/s | Ops: 241.3 ± 0.4 ops/s. Best: 241.7 ops/s |
| node_table_mld | Ops: 58.9 ± 0.6 ops/s. Best: 59.6 ops/s | Ops: 58.5 ± 0.1 ops/s. Best: 58.8 ops/s |
| node_trip_ch | Ops: 298.8 ± 1.6 ops/s. Best: 301.1 ops/s | Ops: 279.4 ± 2.9 ops/s. Best: 285.2 ops/s |
| node_trip_mld | Ops: 99.8 ± 1.2 ops/s. Best: 101.3 ops/s | Ops: 98.1 ± 1.0 ops/s. Best: 99.4 ops/s |
| osrm_contract | Time: 100.34s Peak RAM: 194.83MB | Time: 100.97s Peak RAM: 195.90MB |
| osrm_customize | Time: 1.30s Peak RAM: 117.68MB | Time: 1.30s Peak RAM: 117.72MB |
| osrm_extract | Time: 11.74s Peak RAM: 425.02MB | Time: 11.93s Peak RAM: 429.53MB |
| osrm_partition | Time: 2.11s Peak RAM: 140.69MB | Time: 2.19s Peak RAM: 141.07MB |
| packedvector | random write:<br/>std::vector 11238.7 ms<br/>util::packed_vector 74141.4 ms<br/>slowdown: 6.597<br/>random read:<br/>std::vector 11123.5 ms<br/>util::packed_vector 30804.5 ms<br/>slowdown: 2.76932 | random write:<br/>std::vector 11583.9 ms<br/>util::packed_vector 74065.6 ms<br/>slowdown: 6.39385<br/>random read:<br/>std::vector 11405.6 ms<br/>util::packed_vector 30940.3 ms<br/>slowdown: 2.71273 |
| random_match_ch | 500 matches, default radius<br/>ops: 207.24 ± 0.74 ops/s. best: 208.29ops/s.<br/>total: 275.05 ± 1.06ms. best: 273.65ms.<br/>avg: 4.83 ± 0.02ms<br/>min: 0.13 ± 0.01ms<br/>max: 24.67 ± 0.08ms<br/>p99: 24.67 ± 0.08ms<br/><br/>500 matches, radius=10<br/>ops: 61.39 ± 0.23 ops/s. best: 61.72ops/s.<br/>total: 1042.46 ± 4.05ms. best: 1036.93ms.<br/>avg: 16.29 ± 0.06ms<br/>min: 0.16 ± 0.00ms<br/>max: 234.70 ± 1.56ms<br/>p99: 234.70 ± 1.56ms<br/><br/>500 matches, radius=20<br/>ops: 14.58 ± 0.03 ops/s. best: 14.65ops/s.<br/>total: 4458.65 ± 10.48ms. best: 4437.67ms.<br/>avg: 68.59 ± 0.16ms<br/>min: 0.30 ± 0.00ms<br/>max: 1197.45 ± 4.75ms<br/>p99: 1197.45 ± 4.75ms | 500 matches, default radius<br/>ops: 207.70 ± 0.56 ops/s. best: 208.50ops/s.<br/>total: 274.44 ± 0.73ms. best: 273.38ms.<br/>avg: 4.81 ± 0.01ms<br/>min: 0.14 ± 0.01ms<br/>max: 24.92 ± 0.07ms<br/>p99: 24.92 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 60.37 ± 0.05 ops/s. best: 60.45ops/s.<br/>total: 1060.12 ± 0.84ms. best: 1058.80ms.<br/>avg: 16.56 ± 0.01ms<br/>min: 0.16 ± 0.00ms<br/>max: 249.88 ± 0.58ms<br/>p99: 249.88 ± 0.58ms<br/><br/>500 matches, radius=20<br/>ops: 14.21 ± 0.02 ops/s. best: 14.24ops/s.<br/>total: 4575.69 ± 6.04ms. best: 4564.20ms.<br/>avg: 70.40 ± 0.09ms<br/>min: 0.31 ± 0.01ms<br/>max: 1296.05 ± 2.00ms<br/>p99: 1296.05 ± 2.00ms |
| random_match_mld | 500 matches, default radius<br/>ops: 301.76 ± 1.86 ops/s. best: 303.24ops/s.<br/>total: 188.90 ± 1.17ms. best: 187.97ms.<br/>avg: 3.31 ± 0.02ms<br/>min: 0.11 ± 0.01ms<br/>max: 18.93 ± 0.07ms<br/>p99: 18.93 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 106.70 ± 0.12 ops/s. best: 106.87ops/s.<br/>total: 599.83 ± 0.68ms. best: 598.88ms.<br/>avg: 9.37 ± 0.01ms<br/>min: 0.14 ± 0.00ms<br/>max: 110.81 ± 0.32ms<br/>p99: 110.81 ± 0.32ms<br/><br/>500 matches, radius=20<br/>ops: 21.57 ± 0.03 ops/s. best: 21.61ops/s.<br/>total: 3013.26 ± 3.82ms. best: 3007.77ms.<br/>avg: 46.36 ± 0.06ms<br/>min: 0.19 ± 0.01ms<br/>max: 585.72 ± 1.74ms<br/>p99: 585.72 ± 1.74ms | 500 matches, default radius<br/>ops: 295.91 ± 1.66 ops/s. best: 297.23ops/s.<br/>total: 192.64 ± 1.09ms. best: 191.77ms.<br/>avg: 3.38 ± 0.02ms<br/>min: 0.12 ± 0.01ms<br/>max: 19.28 ± 0.02ms<br/>p99: 19.28 ± 0.02ms<br/><br/>500 matches, radius=10<br/>ops: 104.73 ± 0.21 ops/s. best: 105.05ops/s.<br/>total: 611.08 ± 1.22ms. best: 609.25ms.<br/>avg: 9.55 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 113.73 ± 1.02ms<br/>p99: 113.73 ± 1.02ms<br/><br/>500 matches, radius=20<br/>ops: 21.30 ± 0.04 ops/s. best: 21.35ops/s.<br/>total: 3052.18 ± 6.38ms. best: 3044.12ms.<br/>avg: 46.96 ± 0.10ms<br/>min: 0.19 ± 0.00ms<br/>max: 595.07 ± 1.97ms<br/>p99: 595.07 ± 1.97ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 22781.17 ± 58.52 ops/s. best: 22848.04ops/s.<br/>total: 438.96 ± 1.13ms. best: 437.67ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.02ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17325.42 ± 41.87 ops/s. best: 17365.82ops/s.<br/>total: 577.19 ± 1.40ms. best: 575.84ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13757.00 ± 11.78 ops/s. best: 13771.34ops/s.<br/>total: 726.90 ± 0.62ms. best: 726.15ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 23289.57 ± 63.68 ops/s. best: 23367.93ops/s.<br/>total: 429.38 ± 1.18ms. best: 427.94ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.02ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17686.30 ± 53.99 ops/s. best: 17810.57ops/s.<br/>total: 565.42 ± 1.72ms. best: 561.46ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13986.89 ± 26.62 ops/s. best: 14042.37ops/s.<br/>total: 714.96 ± 1.32ms. best: 712.13ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.14 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 23550.63 ± 60.36 ops/s. best: 23610.71ops/s.<br/>total: 424.62 ± 1.09ms. best: 423.54ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.02ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17848.21 ± 12.07 ops/s. best: 17877.39ops/s.<br/>total: 560.28 ± 0.38ms. best: 559.37ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14152.45 ± 4.74 ops/s. best: 14162.54ops/s.<br/>total: 706.59 ± 0.24ms. best: 706.09ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 23459.23 ± 38.36 ops/s. best: 23490.95ops/s.<br/>total: 426.27 ± 0.70ms. best: 425.70ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.02ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17748.17 ± 21.79 ops/s. best: 17786.33ops/s.<br/>total: 563.44 ± 0.69ms. best: 562.23ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13995.48 ± 29.31 ops/s. best: 14027.56ops/s.<br/>total: 714.52 ± 1.50ms. best: 712.88ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 521.61 ± 3.57 ops/s. best: 526.65ops/s.<br/>total: 1886.59 ± 12.92ms. best: 1868.42ms.<br/>avg: 1.92 ± 0.01ms<br/>min: 0.32 ± 0.01ms<br/>max: 3.29 ± 0.36ms<br/>p99: 2.76 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 611.77 ± 4.73 ops/s. best: 617.08ops/s.<br/>total: 1634.71 ± 12.69ms. best: 1620.55ms.<br/>avg: 1.63 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.06 ± 0.11ms<br/>p99: 3.36 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1005.19 ± 2.84 ops/s. best: 1008.88ops/s.<br/>total: 978.93 ± 2.77ms. best: 975.34ms.<br/>avg: 0.99 ± 0.00ms<br/>min: 0.24 ± 0.01ms<br/>max: 1.73 ± 0.03ms<br/>p99: 1.42 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1163.23 ± 3.90 ops/s. best: 1169.52ops/s.<br/>total: 859.69 ± 2.88ms. best: 855.05ms.<br/>avg: 0.86 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.56 ± 0.06ms<br/>p99: 1.86 ± 0.01ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 511.52 ± 5.98 ops/s. best: 519.01ops/s.<br/>total: 1924.05 ± 22.69ms. best: 1895.93ms.<br/>avg: 1.96 ± 0.02ms<br/>min: 0.32 ± 0.00ms<br/>max: 3.58 ± 0.50ms<br/>p99: 2.81 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 581.91 ± 1.57 ops/s. best: 584.28ops/s.<br/>total: 1718.51 ± 4.63ms. best: 1711.50ms.<br/>avg: 1.72 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.83 ± 0.24ms<br/>p99: 4.03 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 925.66 ± 12.95 ops/s. best: 947.07ops/s.<br/>total: 1063.30 ± 14.93ms. best: 1038.99ms.<br/>avg: 1.08 ± 0.02ms<br/>min: 0.24 ± 0.00ms<br/>max: 1.78 ± 0.03ms<br/>p99: 1.54 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1067.00 ± 8.41 ops/s. best: 1084.59ops/s.<br/>total: 937.28 ± 7.33ms. best: 922.01ms.<br/>avg: 0.94 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.64 ± 0.16ms<br/>p99: 2.33 ± 0.02ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 246.32 ± 4.78 ops/s. best: 254.60ops/s.<br/>total: 3996.88 ± 77.29ms. best: 3864.95ms.<br/>avg: 4.06 ± 0.08ms<br/>min: 0.32 ± 0.01ms<br/>max: 8.84 ± 0.23ms<br/>p99: 6.85 ± 0.18ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 241.78 ± 3.48 ops/s. best: 246.23ops/s.<br/>total: 4137.09 ± 60.28ms. best: 4061.28ms.<br/>avg: 4.14 ± 0.06ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.83 ± 0.39ms<br/>p99: 8.54 ± 0.15ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 332.86 ± 2.01 ops/s. best: 334.98ops/s.<br/>total: 2956.35 ± 17.87ms. best: 2937.45ms.<br/>avg: 3.00 ± 0.02ms<br/>min: 0.29 ± 0.01ms<br/>max: 7.25 ± 0.04ms<br/>p99: 5.18 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 301.13 ± 1.69 ops/s. best: 302.81ops/s.<br/>total: 3321.01 ± 18.79ms. best: 3302.37ms.<br/>avg: 3.32 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.37 ± 0.45ms<br/>p99: 6.62 ± 0.03ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 251.26 ± 5.50 ops/s. best: 258.19ops/s.<br/>total: 3918.90 ± 95.82ms. best: 3811.15ms.<br/>avg: 3.98 ± 0.10ms<br/>min: 0.31 ± 0.01ms<br/>max: 8.66 ± 0.16ms<br/>p99: 6.68 ± 0.11ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 240.93 ± 0.35 ops/s. best: 241.39ops/s.<br/>total: 4150.59 ± 6.25ms. best: 4142.70ms.<br/>avg: 4.15 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.04 ± 0.45ms<br/>p99: 8.58 ± 0.02ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 324.42 ± 2.36 ops/s. best: 327.75ops/s.<br/>total: 3033.28 ± 22.19ms. best: 3002.25ms.<br/>avg: 3.08 ± 0.02ms<br/>min: 0.28 ± 0.00ms<br/>max: 7.38 ± 0.13ms<br/>p99: 5.30 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 298.30 ± 0.78 ops/s. best: 299.52ops/s.<br/>total: 3352.37 ± 9.25ms. best: 3338.67ms.<br/>avg: 3.35 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.56 ± 0.71ms<br/>p99: 6.74 ± 0.02ms |
…